feat(map): add higher-order lettable map operator
