+++ src/operators/index.ts
@@ -1,0 +1,1 @@
+export { map, OperatorFunction } from './map';
